requirements.txt
paho-mqtt
pyyaml
git+https://github.com/babca/python-gsmmodem
